Added samples to db_manage_commands

+++ flask-myrent-api/myrent_app/pictures/pictures.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from myrent_app import db
 from myrent_app.pictures import pictures_bp
 from myrent_app.models import Picture, Flat, PictureSchema, picture_schema
-from myrent_app.utils import allowed_picture
+from myrent_app.utils import allowed_picture, token_landlord_required
 
 
 @pictures_bp.route('/', methods=['GET', 'POST'])
@@ -20,15 +20,23 @@ def upload_test_file():
 
 
 
-@pictures_bp.route('/flats/<int:flat_id>/pictures', methods=['GET'])
-def get_pictures(flat_id: int):
-    flat = Flat.query.get_or_404(flat_id, description=f'Flat with id {flat_id} not found')
-
-    pictures = PictureSchema(many=True).dump(flat.pictures)
+@pictures_bp.route('/pictures', methods=['GET'])
+def get_pictures():
+    pictures = Picture.query.all()
 
     return jsonify({
         'success': True,
-        'data': pictures
+        'data': PictureSchema(many=True).dump(pictures)
+    })
+
+
+@pictures_bp.route('/flats/<int:flat_id>/pictures', methods=['GET'])
+def get_flat_pictures(flat_id: int):
+    flat = Flat.query.get_or_404(flat_id, description=f'Flat with id {flat_id} not found')
+
+    return jsonify({
+        'success': True,
+        'data': PictureSchema(many=True).dump(flat.pictures)
     })
 
 
@@ -43,10 +51,13 @@ def get_picture(picture_id: int):
     })
 
 
-#only landlord
 @pictures_bp.route('/flats/<int:flat_id>/pictures', methods=['POST'])
-def add_picture(flat_id: int):
+@token_landlord_required
+def add_picture(landlord_id: int, flat_id: int):   
     flat = Flat.query.get_or_404(flat_id, description=f'Flat with id {flat_id} not found')
+
+    if flat.landlord_id != landlord_id:
+        abort(404, description=f'Flat with id {flat_id} not found')
 
     file = request.files['picture']
     description = request.form.get('description')
@@ -79,11 +90,14 @@ def add_picture(flat_id: int):
     }), 201
 
 
-#only landlord
 @pictures_bp.route('/pictures/<int:picture_id>', methods=['DELETE'])
-def delete_picture(picture_id: int):
+@token_landlord_required
+def delete_picture(landlord_id: int, picture_id: int):
     picture = Picture.query.get_or_404(picture_id, 
                                 description=f'Picture with id {picture_id} not found')
+
+    if picture.flat.landlord_id != landlord_id:
+        abort(404, description=f'Picture with id {picture_id} not found')
 
     db.session.delete(picture)
     db.session.commit()
